[mapplauncherd] Check sailjaild for sandboxing of apps. JB#54498 OMP#JOLLA-178
Check before executing if the arguments don't contain sailjaild and the
if the app should be launched in a sandbox. In that case prepend the
arguments with sailjail.

This is not sufficient alone as the apps are launched via silica-qt5
booster which can't execute sailjail so this must be also taken into
account already before arguments reach booster.

Signed-off-by: Tomi Leppänen <tomi.leppanen@jolla.com>